ディレクトリ構造の変更、ドキュメント整備
