+++ .gitignore
@@ -1,2 +1,5 @@
-noimage.png
+venv
 prompts.db
+src/__pycache__
+src/resources/images/noimage.png
+
