+++ requirements.txt
@@ -1,0 +1,2 @@
+PyQt5==5.15.9
+requests==2.31.0
