+++ startup.bat
@@ -1,0 +1,2 @@
+call venv\Scripts\activate.bat
+python src/main.py
